security/jwtauth: move from pkg/ccl/jwtauthccl
Move the JWT authentication package out of the CCL tree into
pkg/security/jwtauth. The CCL tree separation is no longer necessary
since all CockroachDB code is now under the CockroachDB Software
License.

Informs #164337

Release note: None
