+++ pkg/ccl/testccl/authccl/BUILD.bazel
@@ -10,9 +10,9 @@ go_test(
     deps = [
         "//pkg/base",
         "//pkg/ccl",
-        "//pkg/ccl/jwtauthccl",
         "//pkg/ccl/ldapccl",
         "//pkg/security/distinguishedname",
+        "//pkg/security/jwtauth",
         "//pkg/security/securityassets",
         "//pkg/security/securitytest",
         "//pkg/security/username",
