+++ pkg/security/jwtauth/BUILD.bazel
@@ -1,13 +1,13 @@
 load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 
 go_library(
-    name = "jwtauthccl",
+    name = "jwtauth",
     srcs = [
         "authentication_jwt.go",
         "authorization_jwt.go",
         "settings.go",
     ],
-    importpath = "github.com/cockroachdb/cockroach/pkg/ccl/jwtauthccl",
+    importpath = "github.com/cockroachdb/cockroach/pkg/security/jwtauth",
     visibility = ["//visibility:public"],
     deps = [
         "//pkg/security/jwthelper",
@@ -31,8 +31,7 @@ go_library(
 )
 
 go_test(
-    name = "jwtauthccl_test",
-    size = "medium",
+    name = "jwtauth_test",
     srcs = [
         "authentication_jwt_test.go",
         "authorization_jwt_test.go",
@@ -40,7 +39,7 @@ go_test(
         "settings_test.go",
     ],
     data = glob(["testdata/**"]),
-    embed = [":jwtauthccl"],
+    embed = [":jwtauth"],
     deps = [
         "//pkg/base",
         "//pkg/ccl",
