+++ pkg/security/jwtauth/main_test.go
@@ -3,7 +3,7 @@
 // Use of this software is governed by the CockroachDB Software License
 // included in the /LICENSE file.
 
-package jwtauthccl_test
+package jwtauth_test
 
 import (
 	"os"
