+++ pkg/security/jwtauth/testdata/accounts.idp1.com_.well-known_openid-configuration
